Use "StrategyName - SongTitle" in status menu items instead of the tab title
The current implementation, which uses a tab's title in the status menu,
works well for sites like soundcloud which update the tab's title according
to the currently playing song. But for sites that don't update the title,
one can't know the current track title simply by looking at beardedspice's
menu. So, instead of using the tab title, we could use the aforementioned
format for the sites whose track info is available and fall back on the
tab title for other sites.